Using LocomotionMode Enum instead of integer.
Because the path planner library is currently doing arithmetic with its own enum,
we opted to only do a mapping in the orogen from the path planner's locomotion mode
(which is treated as an int) to the locomotion switcher's LocomotionMode enum.

The mapping happens just before the output port is written to.